refactor(video-conferencing-service): added reusability
added reusability for functions, provider for crypto-random-string package, moved business logic to services

gh-307
