+++ services/video-conferencing-service/src/controllers/video-chat-archive.controller.ts
@@ -1,31 +1,22 @@
 import {inject} from '@loopback/core';
-import {del, get, param, HttpErrors, put, requestBody} from '@loopback/rest';
+import {del, get, param, put, requestBody} from '@loopback/rest';
 import {authenticate, STRATEGY} from 'loopback4-authentication';
 import {authorize} from 'loopback4-authorization';
 import {PermissionKeys} from '../enums/permission-keys.enum';
-import {
-  VideoChatInterface,
-  S3TargetOptions,
-  AzureTargetOptions,
-} from '../types';
-import {VideoChatBindings} from '../keys';
+import {S3TargetOptions, AzureTargetOptions} from '../types';
+
 import {
   STATUS_CODE,
   CONTENT_TYPE,
   OPERATION_SECURITY_SPEC,
 } from '@sourceloop/core';
-import {repository} from '@loopback/repository';
-import {VideoChatSessionRepository, AuditLogsRepository} from '../repositories';
-import moment from 'moment';
+import {ChatArchiveService} from '../services/chat-archive.service';
+import {ServiceBindings} from '../keys';
 
 export class VideoChatArchiveController {
   constructor(
-    @inject(VideoChatBindings.VideoChatProvider)
-    private readonly videoChatProvider: VideoChatInterface,
-    @repository(VideoChatSessionRepository)
-    private readonly videoChatSessionRepository: VideoChatSessionRepository,
-    @repository(AuditLogsRepository)
-    private readonly auditLogRepository: AuditLogsRepository,
+    @inject(ServiceBindings.ArchiveChatService)
+    public chatArchiveService: ChatArchiveService,
   ) {}
 
   @authenticate(STRATEGY.BEARER)
@@ -47,23 +38,7 @@ export class VideoChatArchiveController {
     },
   })
   async getArchive(@param.path.string('archiveId') archiveId: string) {
-    const archiveExists = await this.videoChatSessionRepository.findOne({
-      where: {
-        archiveId: archiveId,
-      },
-    });
-    if (!archiveExists) {
-      const errorMessage = 'Archive Not Found';
-      await this.auditLogRepository.create({
-        action: 'archive',
-        actionType: 'getArchive',
-        before: {archiveId},
-        after: {errorMessage: errorMessage},
-        actedAt: moment().format(),
-      });
-      throw new HttpErrors.NotFound(errorMessage);
-    }
-    return this.videoChatProvider.getArchives(archiveId);
+    return this.chatArchiveService.getArchive(archiveId);
   }
 
   @authenticate(STRATEGY.BEARER)
@@ -85,7 +60,7 @@ export class VideoChatArchiveController {
     },
   })
   async getArchives() {
-    return this.videoChatProvider.getArchives(null);
+    return this.chatArchiveService.getArchives();
   }
 
   @authenticate(STRATEGY.BEARER)
@@ -109,23 +84,7 @@ export class VideoChatArchiveController {
   async deleteArchive(
     @param.path.string('archiveId') archiveId: string,
   ): Promise<void> {
-    const archiveExists = await this.videoChatSessionRepository.findOne({
-      where: {
-        archiveId: archiveId,
-      },
-    });
-    if (!archiveExists) {
-      const errorMessage = 'Archive Not Found';
-      await this.auditLogRepository.create({
-        action: 'archive',
-        actionType: 'getArchive',
-        before: {archiveId},
-        after: {errorMessage: errorMessage},
-        actedAt: moment().format(),
-      });
-      throw new HttpErrors.NotFound(errorMessage);
-    }
-    return this.videoChatProvider.deleteArchive(archiveId);
+    return this.chatArchiveService.deleteArchive(archiveId);
   }
 
   @authenticate(STRATEGY.BEARER)
@@ -149,17 +108,7 @@ export class VideoChatArchiveController {
   async setUploadTarget(
     @requestBody() body: S3TargetOptions | AzureTargetOptions,
   ): Promise<void> {
-    const {accessKey, secretKey, bucket} = body as S3TargetOptions;
-    const {accountName, accountKey, container} = body as AzureTargetOptions;
-    if (
-      !(accessKey && secretKey && bucket) &&
-      !(accountName && accountKey && container)
-    ) {
-      throw new HttpErrors.BadRequest(
-        'Missing s3/azure credentials. Please check request body',
-      );
-    }
-    await this.videoChatProvider.setUploadTarget(body);
+    return this.chatArchiveService.setUploadTarget(body);
   }
   // TO-DO: will do modifications later
   // @authenticate(STRATEGY.BEARER)
