+++ services/video-conferencing-service/src/providers/vonage/vonage.provider.ts
@@ -14,17 +14,13 @@ import {
 } from './types';
 
 import {VonageService} from './vonage.service';
-import {AuditLogsRepository} from '../../repositories';
-import {repository} from '@loopback/repository';
-import moment from 'moment';
+
 import {HttpErrors} from '@loopback/rest';
 
 export class VonageProvider implements Provider<VonageVideoChat> {
   constructor(
     @service(VonageService)
     private readonly vonageService: VonageService,
-    @repository(AuditLogsRepository)
-    private readonly auditLogRepository: AuditLogsRepository,
   ) {}
   value() {
     return {
@@ -32,25 +28,8 @@ export class VonageProvider implements Provider<VonageVideoChat> {
         meetingOptions: VonageMeetingOptions,
       ): Promise<VonageMeetingResponse> => {
         try {
-          const response = await this.vonageService.getMeetingLink(
-            meetingOptions,
-          );
-          await this.auditLogRepository.create({
-            action: 'session',
-            actionType: 'create-session',
-            before: meetingOptions,
-            after: response,
-            actedAt: moment().format(),
-          });
-          return response;
+          return await this.vonageService.getMeetingLink(meetingOptions);
         } catch (error) {
-          await this.auditLogRepository.create({
-            action: 'session',
-            actionType: 'create-session',
-            before: meetingOptions,
-            after: {errorStack: error.stack},
-            actedAt: moment().format(),
-          });
           throw new HttpErrors.InternalServerError('Error creating session');
         }
       },
@@ -59,36 +38,8 @@ export class VonageProvider implements Provider<VonageVideoChat> {
         options: VonageSessionOptions,
       ): Promise<SessionResponse> => {
         try {
-          const response = await this.vonageService.getToken(
-            sessionId,
-            options,
-          );
-          await this.auditLogRepository.create({
-            action: 'session',
-            actionType: 'get-token',
-            before: {
-              sessionId,
-              ...options,
-            },
-            after: {
-              token: response.token,
-            },
-            actedAt: moment().format(),
-          });
-          return response;
+          return await this.vonageService.getToken(sessionId, options);
         } catch (error) {
-          await this.auditLogRepository.create({
-            action: 'session',
-            actionType: 'get-token',
-            before: {
-              sessionId,
-              ...options,
-            },
-            after: {
-              errorStack: error.stack,
-            },
-            actedAt: moment().format(),
-          });
           throw new HttpErrors.InternalServerError(
             'Error occured while generating Token',
           );
@@ -98,23 +49,8 @@ export class VonageProvider implements Provider<VonageVideoChat> {
         archiveId: string | null,
       ): Promise<ArchiveResponse | ArchiveResponseList> => {
         try {
-          const response = await this.vonageService.getArchives(archiveId);
-          await this.auditLogRepository.create({
-            action: 'archive',
-            actionType: archiveId ? 'getArchive' : 'getArchives',
-            before: archiveId ? {archiveId} : {},
-            after: response,
-            actedAt: moment().format(),
-          });
-          return response;
+          return await this.vonageService.getArchives(archiveId);
         } catch (error) {
-          await this.auditLogRepository.create({
-            action: 'archive',
-            actionType: archiveId ? 'getArchive' : 'getArchives',
-            before: archiveId ? {archiveId} : {},
-            after: {errorStack: error.stack},
-            actedAt: moment().format(),
-          });
           throw new HttpErrors.InternalServerError(
             'Error occured while fetching archive(s)',
           );
@@ -123,25 +59,7 @@ export class VonageProvider implements Provider<VonageVideoChat> {
       deleteArchive: async (archiveId: string) => {
         try {
           await this.vonageService.deleteArchive(archiveId);
-          await this.auditLogRepository.create({
-            action: 'archive',
-            actionType: 'delete-archive',
-            before: {
-              archiveId,
-            },
-            after: {response: 'Archive Deletion Successful!'},
-            actedAt: moment().format(),
-          });
         } catch (error) {
-          await this.auditLogRepository.create({
-            action: 'archive',
-            actionType: 'delete-archive',
-            before: {
-              archiveId,
-            },
-            after: {errorStack: error.stack},
-            actedAt: moment().format(),
-          });
           throw new HttpErrors.InternalServerError(
             'Error occured while deleting an archive',
           );
@@ -150,25 +68,9 @@ export class VonageProvider implements Provider<VonageVideoChat> {
       setUploadTarget: async (
         storageConfig: VonageS3TargetOptions | VonageAzureTargetOptions,
       ): Promise<void> => {
-        const auditLogAction = 'archive';
-        const auditLogActionType = 'set-storage-target';
         try {
           await this.vonageService.setUploadTarget(storageConfig);
-          await this.auditLogRepository.create({
-            action: auditLogAction,
-            actionType: auditLogActionType,
-            before: storageConfig,
-            after: {response: 'Storage Target Success'},
-            actedAt: moment().format(),
-          });
         } catch (error) {
-          await this.auditLogRepository.create({
-            action: auditLogAction,
-            actionType: auditLogActionType,
-            before: storageConfig,
-            after: {errorStack: error.stack},
-            actedAt: moment().format(),
-          });
           throw new HttpErrors.InternalServerError(
             'Error Occured while setting storage target',
           );
