+++ services/video-conferencing-service/tsconfig.json
@@ -2,6 +2,7 @@
   "$schema": "http://json.schemastore.org/tsconfig",
   "extends": "@loopback/build/config/tsconfig.common.json",
   "compilerOptions": {
+   "experimentalDecorators":true,
     "emitDecoratorMetadata": true,
     "outDir": "dist",
     "rootDir": "src"
